src/chomp-common/requirements.txt
PyYAML==3.12
redis==2.10.6
requests==2.13.0
